// face-check-ui/src/app/services/models/payment-history-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface PaymentHistoryRequest {
  amount: number;
  notes: string;
  paymentDate: string;
  paymentType: 'PAYROLL_TAX_941' | 'UNEMPLOYMENT_TAX_940' | 'STATE_UNEMPLOYMENT_TAX' | 'PERSONAL_INSURANCE' | 'WC_Payment' | 'MCTMT_PREPAYMENT' | 'MCTMT_CREDIT';
  quarter: number;
  year: number;
}